+++ CHANGELOG.md
@@ -14,8 +14,10 @@ For information on changes in released versions of Teku, see the [releases page]
 ## Unreleased Changes
 
 ### Breaking Changes
+ - If you have `--Xdata-storage-non-canonical-blocks-enabled` set, this option has been renamed to `--data-storage-non-canonical-blocks-enabled`.
 
 ### Additions and Improvements
+ - Scheduled Altair upgrade on the Prater testnet at epoch 36660.
  - Logged a message to indicate when the node starts and finishes the sync.
  - Upgraded jdk16 and default docker image to use eclipse-tumerin builds of OpenJDK.
  - jdk14 and jdk15 docker images have been upgraded to use the latest Ubuntu. Note that these images will be removed in future versions.
@@ -31,4 +33,5 @@ milestones. Existing Sign type `block` is backward compatible with phase0.
  - Fixed `IllegalStateException: New response submitted after closing AsyncResponseProcessor` errors.
  - Get validator from state should return `404` code rather than a `400` code.
  - Produce attestation data (`/eth/v1/validator/attestation_data`) should return `400` error for future slots, rather than a `500`.
+ - Fixed command-line option `--Xdata-storage-non-canonical-blocks-enabled` which was marked as a development option (-X) but not hidden.